requirements.txt
allure-pytest==2.8.6
allure-python-commons==2.8.6
atomicwrites==1.3.0
attrs==19.3.0
certifi==2020.12.5
chardet==3.0.4
colorama==0.4.1
configparser==4.0.2
crayons==0.3.0
idna==2.8
importlib-metadata==0.23
iniconfig==1.1.1
more-itertools==7.2.0
msedge-selenium-tools==3.141.3
packaging==19.2
pluggy==0.13.0
py==1.10.0
pyparsing==2.4.4
pytest==5.2.2
requests==2.22.0
selenium==3.141.0
six==1.13.0
toml==0.10.2
urllib3==1.25.9
wcwidth==0.1.7
webdriver-manager==3.4.0
xlrd==1.2.0
zipp==0.6.0
pytest-yaml
loguru